requirements.txt
azure-storage-blob==1.4.0
azure-cosmosdb-table==1.0.5
pyyaml==3.13
python-dotenv==0.10.1
Pillow==5.4.1
python-dotenv==0.10.1
webp-converter==4.0.1
sortedcontainers==2.1.0